+++ autoprotocol/container.py
@@ -228,14 +228,20 @@ class WellGroup(object):
 
         Parameters
         ----------
-        other : WellGroup
+        other : WellGroup or list of Wells
             WellGroup to extend this WellGroup.
 
         """
-        if not isinstance(other, WellGroup):
+        if not isinstance(other, (WellGroup,list)):
             raise TypeError("Input given is not of type 'WellGroup'.")
         else:
-            return self.wells.extend(other.wells)
+            if isinstance(other, list):
+                for x in other:
+                    if type(x) != Well:
+                        raise TypeError("Input given is not of type 'Well'.")
+                return self.wells.extend(WellGroup(other).wells)
+            else:
+                return self.wells.extend(other.wells)
 
     def set_group_name(self, name):
         """
@@ -249,13 +255,6 @@ class WellGroup(object):
         self.name = name
         return self
 
-    def get_well_list(self):
-        """
-        Return a numerical list (robot indices) of what wells have liquid in them.
-
-        """
-        return [x.index for x in self.wells]
-
     def wells_with(self, prop, val=None):
         """
         Returns a wellgroup of wells with the specified property and value
@@ -268,9 +267,9 @@ class WellGroup(object):
             the value assigned to the property
         """
         assert type(prop) is str, "property is not a string: %r" % prop
-        assert type(val) is str, "value is not a string: %r" % val
+        assert type(val) is None or str, "value is not a string: %r" % val
         if val:
-            return WellGroup([w for w in self.wells if prop in w.properties and w.properties[pop] is val])
+            return WellGroup([w for w in self.wells if prop in w.properties and w.properties[prop] is val])
         else:
             return WellGroup([w for w in self.wells if prop in w.properties])
 
@@ -304,6 +303,18 @@ class WellGroup(object):
         else:
             self.wells = self.wells[:i] + [well] + self.wells[i:]
 
+    def __setitem__(self, key, item):
+        """
+        Set a specific Well in a WellGroup.
+
+        Parameters
+        ----------
+        key : int
+            Well reference in robotized form.
+        item: Well
+            Well or WellGroup to be added
+        """
+        self.wells[key] = item
 
     def __getitem__(self, key):
         """