+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 Added
 -is_resource_id added to dispense and dispense_full_plate instructions
 -dispense instruction tests
+-autocover before incubate
+-assertions and tests for flow_analyze
 
 Changed
 -dispense instruction json outputs
@@ -14,6 +16,7 @@ Changed
 Removed
 
 Fixed
+-spin auto-cover
 
 ## Unreleased
 ---