+++ CHANGELOG.md
@@ -54,6 +54,8 @@ Removed
 
 Fixed
 
+Fixed
+
 ## v3.7.0 - 2016-06-14
 ---
 Added